Move CSS injection tests to plt_css_test.js.
